requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
dnspython==2.7.0
dotenv==0.9.9
gunicorn==23.0.0
mongoengine==0.29.1
packaging==25.0
python-dotenv==1.1.1
pytz==2025.2
sqlparse==0.5.3
pymongo==3.11.4
tzdata==2025.2
whitenoise==6.9.0